move SegmentUnionFind to the one file where it's used
