+++ find_duplicates.py
@@ -1,9 +1,57 @@
 import numpy
 
-from utils import SegmentUnionFind
-
 
 _MAX_TOKEN_CHAIN = 100  # Sequences at least this long get the most extreme hue
+
+
+class SegmentUnionFind:
+    """
+    UnionFind is sometimes named DisjointSet. Our data structure is different
+    from the usual one because it represents a set of duplicated tokens, with a
+    top-right and bottom-left coordinate, in addition to the size.
+    """
+    def __init__(self, r, c):
+        self._size = 1
+        self._root = self
+        self.top = (r, c)
+        self.bottom = (r, c)
+
+    def get_root(self):
+        if self._root is self:
+            return self
+
+        self._root = self._root.get_root()
+        return self._root
+
+    def size(self):
+        return self.get_root()._size
+
+    def merge(self, other):
+        if self.size() > other.size():
+            large_root = self.get_root()
+            small_root = other.get_root()
+        else:
+            large_root = other.get_root()
+            small_root = self.get_root()
+
+        large_root._size += small_root.size()
+        small_root._root = large_root
+
+        if sum(small_root.top) < sum(large_root.top):
+            # The top of the small section is further towards the top-left
+            # corner. Use it as the new top.
+            large_root.top = small_root.top
+
+        if sum(small_root.bottom) > sum(large_root.bottom):
+            # The bottom of the small section is further towards the
+            # bottom-right corner. Use it as the new bottom.
+            large_root.bottom = small_root.bottom
+
+    def __str__(self):
+        root = self.get_root()
+        return (f"(Segment size {root.size()} "
+                f"from ({root.top_r}, {root.top_c}) "
+                f"to ({root.bottom_r}, {root.bottom_c}))")
 
 
 def get_lengths(matrix, is_single_file):
